+++ db/migrate/20150922131221_rename_column_to_last_activity.rb
@@ -1,0 +1,5 @@
+class RenameColumnToLastActivity < ActiveRecord::Migration
+  def change
+    rename_column :servers, :last_test_run_date, :last_activity_date
+  end
+end