+++ db/migrate/20150922145945_add_server_running_test_column.rb
@@ -1,0 +1,5 @@
+class AddServerRunningTestColumn < ActiveRecord::Migration
+  def change
+    add_column :servers, :executing_command_now, :boolean, default: false
+  end
+end