+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 * Fix collapsing navbar on mobile devices (small screens)
 * Add ability to get last test run date for each server
+* Servers will turn off after specific amount of time of inactivity
 
 ## 1.4.2
 * Remove executed containers, that no need. They took a lot of space